requirements/common-requirements.txt
chardet==2.1.1
colorama==0.2.7
numpy==1.8.0
PyYaml==3.10